Add table of contents and heading links (#26)
* Currently section headings aren't linkable
* We follow this
[recipe](https://tekki-tipps.de/en/hugo-header-anchor-links/) to create
a "hyperlink" icon which can be used to get a link to the header.
* We also add "toc:true" to the frontmatter of articles in order to add
a table of contents to a given page

  * This is styled according to the partial

https://github.com/jlewi/roboweb-docs/blob/4f45795d001290c205b793813d8df8d69df8bf07/roboweb/themes/ananke/layouts/partials/menu-contextual.html#L6

* We update the value "whatsthis" in roboweb/themes/ananke/i18n/en.toml
because otherwise the title of the table of contents looks funky

* I tried to add a table of contents to our main page corresponding to
contents/_index.md but I couldn't get it to render on the side. This
uses the "index.html" layout page. It kept rendering at the bottom of
the page which didn't look good.